// src/app/modules/result/models/examResultSubjectWiseDtl.ts
export class ResultSubjectWiseDetail {
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    GROUP_CODE: string;
    SHIFT_CODE: string;
    VERSION_CODE: string;
    SESSION_CODE: string;
    SECTION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    STUDENT_CODE: string[];
    STUDENT_ID: string;
    SUBJECT_TYPE: number;
    PROCESS_ID: number;
    SUBJECT_CODE: any;
    SUBJECT_NAME: string;
    EXAM_CODE: string;
    STUDENT_NAME: string[];
    EXAM_TYPE: number[];
    MARK: any[];
    TERM_ID: number;
    USER_CODE: string;

    EXAM_1_MARK: number[];
    EXAM_2_MARK: number[];
    EXAM_3_MARK: number[];
    EXAM_4_MARK: number[];
    EXAM_5_MARK: number[];
    EXAM_6_MARK: number[];

    // T2_EXAM_1_MARK: number[];
    // T2_EXAM_2_MARK: number[];
    // T2_EXAM_3_MARK: number[];
    // T2_EXAM_4_MARK: number[];
    // T2_EXAM_5_MARK: number[];
    // T2_EXAM_6_MARK: number[];

    // T3_EXAM_1_MARK: number[];
    // T3_EXAM_2_MARK: number[];
    // T3_EXAM_3_MARK: number[];
    // T3_EXAM_4_MARK: number[];
    // T3_EXAM_5_MARK: number[];
    // T3_EXAM_6_MARK: number[];
    
    // T4_EXAM_1_MARK: number[];
    // T4_EXAM_2_MARK: number[];
    // T4_EXAM_3_MARK: number[];
    // T4_EXAM_4_MARK: number[];
    // T4_EXAM_5_MARK: number[];
    // T4_EXAM_6_MARK: number[];
    
    User_Name: string;
}




export class ResultSubjectWiseDetailVM {
    ORG_CODE: string;
    CAMPUS_CODE: string;
    CLASS_CODE: string;
    GROUP_CODE: string;
    SHIFT_CODE: string;
    VERSION_CODE: string;
    SESSION_CODE: string;
    SECTION_CODE: string;
    YEAR_CODE: string;
    SEMESTER_CODE: string;
    STUDENT_CODE: string[];
    STUDENT_ID: string;
    SUBJECT_TYPE: number;
    PROCESS_ID: number;
    SUBJECT_CODE: string;
    EXAM_CODE: string;
    STUDENT_NAME: string[];
    EXAM_TYPE_ID: number[];
    MARK: number[];
    TERM_ID: number;
    EXAM_TYPE: string;

    EXAM_1_MARK: number[];
    EXAM_2_MARK: number[];
    EXAM_3_MARK: number[];
    EXAM_4_MARK: number[];
    EXAM_5_MARK: number[];
    EXAM_6_MARK: number[];
    // T1_EXAM_1_MARK: number[];
    // T1_EXAM_2_MARK: number[];
    // T1_EXAM_3_MARK: number[];
    // T1_EXAM_4_MARK: number[];
    // T1_EXAM_5_MARK: number[];
    // T1_EXAM_6_MARK: number[];

    // T2_EXAM_1_MARK: number[];
    // T2_EXAM_2_MARK: number[];
    // T2_EXAM_3_MARK: number[];
    // T2_EXAM_4_MARK: number[];
    // T2_EXAM_5_MARK: number[];
    // T2_EXAM_6_MARK: number[];

    // T3_EXAM_1_MARK: number[];
    // T3_EXAM_2_MARK: number[];
    // T3_EXAM_3_MARK: number[];
    // T3_EXAM_4_MARK: number[];
    // T3_EXAM_5_MARK: number[];
    // T3_EXAM_6_MARK: number[];
    
    // T4_EXAM_1_MARK: number[];
    // T4_EXAM_2_MARK: number[];
    // T4_EXAM_3_MARK: number[];
    // T4_EXAM_4_MARK: number[];
    // T4_EXAM_5_MARK: number[];
    // T4_EXAM_6_MARK: number[];
    
    User_Name: string;
}